fix: Tira a responsabilidade do Model e passa para dentro do DTO , usando streams no mapper

+++ src/main/java/dev/java10x/CadastroDeNinjas/Ninjas/NinjaService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class NinjaService {
@@ -18,14 +19,17 @@ public class NinjaService {
     }
 
     //Listar todos os meus ninjas
-    public List<NinjaModel> listarNinjas() {
-        return ninjaRepository.findAll();
+    public List<NinjaDTO> listarNinjas() {
+        List<NinjaModel> ninjas = ninjaRepository.findAll();
+        return ninjas.stream()
+                .map(ninjaMapper::map)
+                .collect(Collectors.toList());
     }
 
     //Lista os ninjas por ID
-    public NinjaModel listaNinjasPorID(Long ID){
+    public NinjaDTO listaNinjasPorID(Long ID){
         Optional<NinjaModel> ninja =  ninjaRepository.findById(ID);
-        return ninja.orElse(null);
+        return ninja.map(ninjaMapper::map).orElse(null);
     }
 
     //Criar um novo ninja
@@ -38,16 +42,22 @@ public class NinjaService {
 
     //Deletar o ninja - Tem que ser um método void
     public void deletarNinjaPorId(Long id){
-        ninjaRepository.deleteById(id);
+        Optional<NinjaModel> ninjaDeletado = ninjaRepository.findById(id);
+        if (ninjaDeletado.isPresent()) {
+            ninjaRepository.deleteById(id);
+        }
     }
 
     //Atualizar Ninja
-    public NinjaModel atualizarNinja(Long id, NinjaModel ninjaAtualizado){
-        if (ninjaRepository.existsById(id)){
+    public NinjaDTO atualizarNinja(Long id, NinjaDTO ninjaDTO){
+        Optional<NinjaModel> ninjaExistente =  ninjaRepository.findById(id);
+
+        if (ninjaExistente.isPresent()) {
+            NinjaModel ninjaAtualizado = ninjaMapper.map(ninjaDTO);
             ninjaAtualizado.setId(id);
-            return ninjaRepository.save(ninjaAtualizado);
-        }
-        return null;
+            NinjaModel ninjaSalvo = ninjaRepository.save(ninjaAtualizado);
+            return ninjaMapper.map(ninjaSalvo);
+        } return null;
     }
 
 }
